+++ app/src/androidTest/java/woowacourse/movie/ui/home/TheaterSelectionBottomSheetFragmentTest.kt
@@ -6,6 +6,7 @@ import androidx.test.ext.junit.runners.AndroidJUnit4
 import org.junit.Test
 import org.junit.jupiter.api.Assertions.assertEquals
 import org.junit.runner.RunWith
+import woowacourse.movie.ui.home.theater.TheaterSelectionBottomSheetFragment
 
 @RunWith(AndroidJUnit4::class)
 class TheaterSelectionBottomSheetFragmentTest {
