+++ app/src/main/java/woowacourse/movie/ui/home/theater/TheaterSelectionBottomSheetFragment.kt
@@ -1,4 +1,4 @@
-package woowacourse.movie.ui.home
+package woowacourse.movie.ui.home.theater
 
 import android.content.Intent
 import android.os.Bundle
@@ -12,6 +12,7 @@ import woowacourse.movie.databinding.FragmentTheaterSelectionBottomSheetBinding
 import woowacourse.movie.model.data.MovieContentsImpl
 import woowacourse.movie.model.data.TheatersImpl
 import woowacourse.movie.model.movie.Theater
+import woowacourse.movie.ui.home.MovieHomeKey
 import woowacourse.movie.ui.home.adapter.TheaterAdapter
 import woowacourse.movie.ui.reservation.MovieReservationActivity
