+++ app/src/main/java/woowacourse/movie/ui/home/theater/TheaterSelectionContract.kt
@@ -1,4 +1,4 @@
-package woowacourse.movie.ui.home
+package woowacourse.movie.ui.home.theater
 
 import woowacourse.movie.model.movie.Theater
